renamed grad_p to momentum_energy
