+++ include/sph/math.hpp
